+++ src/model/guild/guild_id.rs
@@ -1238,7 +1238,7 @@ impl GuildId {
     /// [`utils::shard_id`]: crate::utils::shard_id
     #[cfg(all(feature = "cache", feature = "utils"))]
     #[inline]
-    pub fn shard_id(self, cache: impl AsRef<Cache>) -> u64 {
+    pub fn shard_id(self, cache: impl AsRef<Cache>) -> u32 {
         crate::utils::shard_id(self.0, cache.as_ref().shard_count())
     }
